Split store into two separate features
