+++ apps/kg-explorer/src/app/state/filters.store.ts
@@ -1,4 +1,5 @@
 import { signalStore } from '@ngrx/signals';
+import { withDigitalObjectsData } from './with-do-data.feature';
 import { withFilters } from './with-filters.feature';
 
-export const FiltersStore = signalStore({ providedIn: 'root' }, withFilters());
+export const FiltersStore = signalStore({ providedIn: 'root' }, withDigitalObjectsData(), withFilters());
